Extract Expansion logic out of RichTextCell
