+++ src/scenes/Dashboard/ProgressReportsWidget/RichTextCell.tsx
@@ -1,54 +1,26 @@
 import { Box, Typography } from '@mui/material';
-import {
-  GridState,
-  GridRenderCellParams as RenderCellParams,
-  useGridSelector,
-} from '@mui/x-data-grid';
-import { extendSx, RichTextJson, StyleProps } from '~/common';
+import { GridRenderCellParams as RenderCellParams } from '@mui/x-data-grid';
+import { RichTextJson, StyleProps } from '~/common';
 import {
   BlockProps,
   Renderers,
   RichTextView,
   Text,
 } from '../../../components/RichText';
+import { ExpansionCell } from './ExpansionCell';
 import { ProgressReportsDataGridRowFragment as ProgressReport } from './progressReportsDataGridRow.graphql';
 
 type CellParams = RenderCellParams<ProgressReport, RichTextJson>;
 
-export const RichTextCell = ({
-  id,
-  value,
-  api,
-  sx,
-  className,
-}: Pick<CellParams, 'value' | 'id' | 'api'> & StyleProps) => {
-  const selectedRows = useGridSelector(
-    { current: api },
-    (state: GridState) => state.rowSelection
-  );
-  const isExpanded = selectedRows.includes(id);
-
-  if (!value) return null;
+export const RichTextCell = (
+  props: Pick<CellParams, 'value' | 'id' | 'api'> & StyleProps
+) => {
+  if (!props.value) return null;
 
   return (
-    <Box
-      sx={[
-        {
-          overflow: 'hidden',
-          textWrap: 'wrap',
-          display: isExpanded ? 'contents' : '-webkit-box',
-          WebkitLineClamp: '2',
-          WebkitBoxOrient: 'vertical',
-
-          // No trailing spacing on response
-          '& > *:last-child': { mb: 0 },
-        },
-        ...extendSx(sx),
-      ]}
-      className={className}
-    >
-      <RichTextView data={value} renderers={renderers} />
-    </Box>
+    <ExpansionCell {...props}>
+      <RichTextView data={props.value} renderers={renderers} />
+    </ExpansionCell>
   );
 };
